Move zink to vulkan drivers
While technically the zink driver is a dri driver it
depends on vulkan drivers to be available and it won't
work without them, it also doesn't depend on any of the
dri drivers. The dri and vulkan driver packages are
independent of one another and there are use cases to
remote one or the other due to size.