Move dri_gbm.so to mesa-dri-drivers
dri_gbm.so is linked to libgallium and uses the other DRI drivers, so the
only way to make mesa-dri-drivers optional is to put them all together.
This allows a minimal libgbm to be included in the flatpak runtime while
the actual mesa implementations (with all their dependencies) can be split
out into a runtime extension.