Avoid devel-to-runtime dependencies in flatpak SDKs
We're moving towards a system for flatpaks where the mesa implementation
will be provided as an extension that is automatically loaded into the
flatpak runtimes and SDKs (where libgbm, libglvnd, libva, libvulkan, etc.
remain therein to load the implemenations).

This is already done for the runtimes, but for the SDKs, the mesa -devel
packages need to not pull in the runtimes.  With libglvnd providing the
external APIs, none of the mesa-*-devel packages provide an external link
interface, just a header interface and pkg-config data.

Since these -devel packages already require libglvnd-devel, which requires
libglvnd-egl and libglvnd-glx, which (normally) pull in mesa-libEGL and
mesa-libGL respectively, all that is needed to lock in a matching version.
This way, when building for flatpaks which intentionally do not do so (as
above), they are not pulled into the SDK either.